requirements.txt
python-dotenv>=1.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
openai==1.51.0
sentence-transformers==3.1.1
faiss-cpu==1.9.0
numpy==1.26.4
httpx==0.27.2
slowapi==0.1.9
python-multipart==0.0.12
requests>=2.31.0
